backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.4
pydantic-settings==2.2.1
langgraph==0.0.62
chromadb==0.5.3
pdfplumber==0.11.0
scipy==1.12.0
anthropic==0.30.0
groq==0.9.0
python-dotenv==1.0.1
websockets==12.0
pytest==8.1.1
httpx==0.27.0

